requirements/requirements-default.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make requirements
#
ansi2html==1.9.1
    # via jupyter-dash
asttokens==2.4.1
    # via stack-data
backcall==0.2.0
    # via ipython
blinker==1.7.0
    # via flask
causallib==0.9.6
    # via -r requirements/requirements-default.in
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   flask
    #   nltk
comm==0.2.2
    # via
    #   ipykernel
    #   ipywidgets
contourpy==1.1.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dash==2.9.2
    # via
    #   -r requirements/requirements-default.in
    #   dash-bootstrap-components
    #   dash-daq
    #   jupyter-dash
dash-bootstrap-components==1.4.1
    # via -r requirements/requirements-default.in
dash-core-components==2.0.0
    # via
    #   -r requirements/requirements-default.in
    #   dash
dash-daq==0.5.0
    # via -r requirements/requirements-default.in
dash-html-components==2.0.0
    # via
    #   -r requirements/requirements-default.in
    #   dash
dash-table==5.0.0
    # via dash
debugpy==1.8.1
    # via ipykernel
decorator==5.1.1
    # via ipython
dgl==1.1.2
    # via -r requirements/requirements-default.in
executing==2.0.1
    # via stack-data
filelock==3.13.3
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
flask==3.0.2
    # via
    #   dash
    #   jupyter-dash
fonttools==4.50.0
    # via matplotlib
fsspec==2024.3.1
    # via
    #   huggingface-hub
    #   torch
gravis==0.1.0
    # via -r requirements/requirements-default.in
huggingface-hub==0.17.3
    # via
    #   tokenizers
    #   transformers
idna==3.6
    # via requests
imbalanced-learn==0.12.0
    # via imblearn
imblearn==0.0
    # via -r requirements/requirements-default.in
importlib-metadata==7.1.0
    # via
    #   flask
    #   jupyter-client
    #   numba
importlib-resources==6.4.0
    # via matplotlib
ipykernel==6.29.3
    # via jupyter-dash
ipython==8.12.3
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-dash
ipywidgets==8.1.2
    # via -r requirements/requirements-default.in
itsdangerous==2.1.2
    # via flask
jedi==0.19.1
    # via ipython
jinja2==3.1.3
    # via
    #   flask
    #   torch
joblib==1.3.2
    # via
    #   imbalanced-learn
    #   nltk
    #   pynndescent
    #   scikit-learn
jupyter-client==8.6.1
    # via ipykernel
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
jupyter-dash==0.4.2
    # via -r requirements/requirements-default.in
jupyterlab-widgets==3.0.10
    # via ipywidgets
kiwisolver==1.4.5
    # via matplotlib
llvmlite==0.41.1
    # via
    #   numba
    #   pynndescent
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.7.3
    # via
    #   -r requirements/requirements-default.in
    #   causallib
    #   seaborn
    #   wordcloud
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mpmath==1.3.0
    # via sympy
nest-asyncio==1.6.0
    # via
    #   ipykernel
    #   jupyter-dash
networkx==3.1
    # via
    #   causallib
    #   dgl
    #   torch
nltk==3.8.1
    # via -r requirements/requirements-default.in
numba==0.58.1
    # via
    #   pynndescent
    #   umap-learn
numpy==1.24.4
    # via
    #   -r requirements/requirements-default.in
    #   causallib
    #   contourpy
    #   dgl
    #   imbalanced-learn
    #   matplotlib
    #   numba
    #   pandas
    #   patsy
    #   scikit-learn
    #   scipy
    #   seaborn
    #   statsmodels
    #   transformers
    #   umap-learn
    #   wordcloud
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==8.9.2.26
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.18.1
    # via torch
nvidia-nvjitlink-cu12==12.4.99
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
packaging==24.0
    # via
    #   huggingface-hub
    #   ipykernel
    #   matplotlib
    #   plotly
    #   statsmodels
    #   transformers
pandas==2.0.3
    # via
    #   -r requirements/requirements-default.in
    #   causallib
    #   seaborn
    #   statsmodels
parso==0.8.3
    # via jedi
patsy==0.5.6
    # via statsmodels
pexpect==4.9.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==10.2.0
    # via
    #   matplotlib
    #   wordcloud
platformdirs==4.2.0
    # via jupyter-core
plotly==5.14.1
    # via
    #   -r requirements/requirements-default.in
    #   dash
prompt-toolkit==3.0.43
    # via ipython
psutil==5.9.8
    # via
    #   dgl
    #   ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.17.2
    # via ipython
pynndescent==0.5.11
    # via umap-learn
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   transformers
pyzmq==25.1.2
    # via
    #   ipykernel
    #   jupyter-client
regex==2023.12.25
    # via
    #   nltk
    #   transformers
requests==2.31.0
    # via
    #   dgl
    #   huggingface-hub
    #   jupyter-dash
    #   transformers
retrying==1.3.4
    # via jupyter-dash
safetensors==0.4.2
    # via transformers
scikit-learn==1.1.3
    # via
    #   -r requirements/requirements-default.in
    #   causallib
    #   imbalanced-learn
    #   pynndescent
    #   umap-learn
scipy==1.10.1
    # via
    #   causallib
    #   dgl
    #   imbalanced-learn
    #   pynndescent
    #   scikit-learn
    #   statsmodels
    #   umap-learn
seaborn==0.13.0
    # via -r requirements/requirements-default.in
six==1.16.0
    # via
    #   asttokens
    #   patsy
    #   python-dateutil
    #   retrying
stack-data==0.6.3
    # via ipython
statsmodels==0.14.1
    # via causallib
sympy==1.12
    # via torch
tenacity==8.2.3
    # via plotly
threadpoolctl==3.4.0
    # via
    #   imbalanced-learn
    #   scikit-learn
tokenizers==0.14.1
    # via transformers
torch==2.1.0
    # via -r requirements/requirements-default.in
torchsummary==1.5.1
    # via -r requirements/requirements-default.in
tornado==6.4
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.66.1
    # via
    #   -r requirements/requirements-default.in
    #   dgl
    #   huggingface-hub
    #   nltk
    #   transformers
    #   umap-learn
traitlets==5.14.2
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
transformers==4.35.0
    # via -r requirements/requirements-default.in
triton==2.1.0
    # via torch
typing-extensions==4.10.0
    # via
    #   huggingface-hub
    #   ipython
    #   torch
tzdata==2024.1
    # via pandas
umap-learn==0.5.5
    # via -r requirements/requirements-default.in
urllib3==2.2.1
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
werkzeug==3.0.1
    # via flask
widgetsnbextension==4.0.10
    # via ipywidgets
wordcloud==1.9.2
    # via -r requirements/requirements-default.in
zipp==3.18.1
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
